requirements.txt
pytest~=8.3.4
pytest-cov
allure-pytest
allure-python-commons~=2.13.5
selenium~=4.27.1
webdriver-manager~=4.0.2
